+++ layers/Schema/packages/user-state-mutations/src/FieldResolvers/ObjectType/RootObjectTypeFieldResolver.php
@@ -10,18 +10,18 @@ use PoP\ComponentModel\Schema\SchemaTypeModifiers;
 use PoP\ComponentModel\TypeResolvers\ConcreteTypeResolverInterface;
 use PoP\ComponentModel\TypeResolvers\ObjectType\ObjectTypeResolverInterface;
 use PoP\Engine\TypeResolvers\ObjectType\RootObjectTypeResolver;
-use PoP\Engine\TypeResolvers\ScalarType\StringScalarTypeResolver;
 use PoPSchema\Users\TypeResolvers\ObjectType\UserObjectTypeResolver;
-use PoPSchema\UserStateMutations\MutationResolvers\LoginMutationResolver;
+use PoPSchema\UserStateMutations\MutationResolvers\LoginTaggedMutationResolver;
 use PoPSchema\UserStateMutations\MutationResolvers\LogoutMutationResolver;
 use PoPSchema\UserStateMutations\MutationResolvers\MutationInputProperties;
+use PoPSchema\UserStateMutations\TypeResolvers\InputObjectType\LoginCredentialsTaggedInputObjectTypeResolver;
 
 class RootObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResolver
 {
     private ?UserObjectTypeResolver $userObjectTypeResolver = null;
-    private ?LoginMutationResolver $loginMutationResolver = null;
+    private ?LoginTaggedMutationResolver $loginTaggedMutationResolver = null;
     private ?LogoutMutationResolver $logoutMutationResolver = null;
-    private ?StringScalarTypeResolver $stringScalarTypeResolver = null;
+    private ?LoginCredentialsTaggedInputObjectTypeResolver $loginCredentialsTaggedInputObjectTypeResolver = null;
 
     final public function setUserObjectTypeResolver(UserObjectTypeResolver $userObjectTypeResolver): void
     {
@@ -31,13 +31,13 @@ class RootObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResolv
     {
         return $this->userObjectTypeResolver ??= $this->instanceManager->getInstance(UserObjectTypeResolver::class);
     }
-    final public function setLoginMutationResolver(LoginMutationResolver $loginMutationResolver): void
+    final public function setLoginTaggedMutationResolver(LoginTaggedMutationResolver $loginTaggedMutationResolver): void
     {
-        $this->loginMutationResolver = $loginMutationResolver;
+        $this->loginTaggedMutationResolver = $loginTaggedMutationResolver;
     }
-    final protected function getLoginMutationResolver(): LoginMutationResolver
+    final protected function getLoginTaggedMutationResolver(): LoginTaggedMutationResolver
     {
-        return $this->loginMutationResolver ??= $this->instanceManager->getInstance(LoginMutationResolver::class);
+        return $this->loginTaggedMutationResolver ??= $this->instanceManager->getInstance(LoginTaggedMutationResolver::class);
     }
     final public function setLogoutMutationResolver(LogoutMutationResolver $logoutMutationResolver): void
     {
@@ -47,13 +47,13 @@ class RootObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResolv
     {
         return $this->logoutMutationResolver ??= $this->instanceManager->getInstance(LogoutMutationResolver::class);
     }
-    final public function setStringScalarTypeResolver(StringScalarTypeResolver $stringScalarTypeResolver): void
+    final public function setLoginCredentialsTaggedInputObjectTypeResolver(LoginCredentialsTaggedInputObjectTypeResolver $loginCredentialsTaggedInputObjectTypeResolver): void
     {
-        $this->stringScalarTypeResolver = $stringScalarTypeResolver;
+        $this->loginCredentialsTaggedInputObjectTypeResolver = $loginCredentialsTaggedInputObjectTypeResolver;
     }
-    final protected function getStringScalarTypeResolver(): StringScalarTypeResolver
+    final protected function getLoginCredentialsTaggedInputObjectTypeResolver(): LoginCredentialsTaggedInputObjectTypeResolver
     {
-        return $this->stringScalarTypeResolver ??= $this->instanceManager->getInstance(StringScalarTypeResolver::class);
+        return $this->loginCredentialsTaggedInputObjectTypeResolver ??= $this->instanceManager->getInstance(LoginCredentialsTaggedInputObjectTypeResolver::class);
     }
 
     public function getObjectTypeResolverClassesToAttachTo(): array
@@ -84,8 +84,7 @@ class RootObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResolv
     {
         return match ($fieldName) {
             'loginUser' => [
-                MutationInputProperties::USERNAME_OR_EMAIL => $this->getStringScalarTypeResolver(),
-                MutationInputProperties::PASSWORD => $this->getStringScalarTypeResolver(),
+                MutationInputProperties::CREDENTIALS => $this->getLoginCredentialsTaggedInputObjectTypeResolver(),
             ],
             default => parent::getFieldArgNameTypeResolvers($objectTypeResolver, $fieldName),
         };
@@ -94,8 +93,7 @@ class RootObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResolv
     public function getFieldArgDescription(ObjectTypeResolverInterface $objectTypeResolver, string $fieldName, string $fieldArgName): ?string
     {
         return match ([$fieldName => $fieldArgName]) {
-            ['loginUser' => MutationInputProperties::USERNAME_OR_EMAIL] => $this->getTranslationAPI()->__('The username or email', 'user-state-mutations'),
-            ['loginUser' => MutationInputProperties::PASSWORD] => $this->getTranslationAPI()->__('The password', 'user-state-mutations'),
+            ['loginUser' => MutationInputProperties::CREDENTIALS] => $this->getTranslationAPI()->__('Choose which credentials to use to log-in, and provide them', 'user-state-mutations'),
             default => parent::getFieldArgDescription($objectTypeResolver, $fieldName, $fieldArgName),
         };
     }
@@ -103,18 +101,15 @@ class RootObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResolv
     public function getFieldArgTypeModifiers(ObjectTypeResolverInterface $objectTypeResolver, string $fieldName, string $fieldArgName): int
     {
         return match ([$fieldName => $fieldArgName]) {
-            ['loginUser' => MutationInputProperties::USERNAME_OR_EMAIL],
-            ['loginUser' => MutationInputProperties::PASSWORD]
-                => SchemaTypeModifiers::MANDATORY,
-            default
-                => parent::getFieldArgTypeModifiers($objectTypeResolver, $fieldName, $fieldArgName),
+            ['loginUser' => MutationInputProperties::CREDENTIALS] => SchemaTypeModifiers::MANDATORY,
+            default => parent::getFieldArgTypeModifiers($objectTypeResolver, $fieldName, $fieldArgName),
         };
     }
 
     public function getFieldMutationResolver(ObjectTypeResolverInterface $objectTypeResolver, string $fieldName): ?MutationResolverInterface
     {
         return match ($fieldName) {
-            'loginUser' => $this->getLoginMutationResolver(),
+            'loginUser' => $this->getLoginTaggedMutationResolver(),
             'logoutUser' => $this->getLogoutMutationResolver(),
             default => parent::getFieldMutationResolver($objectTypeResolver, $fieldName),
         };